refactor: move MCP tools to separate package and update server
